# docs/swagger.yaml
basePath: /
definitions:
  domain.BackgroundCheckStatus:
    enum:
    - APPROVED
    - PENDING
    - REJECTED
    type: string
    x-enum-varnames:
    - BackgroundCheckStatusApproved
    - BackgroundCheckStatusPending
    - BackgroundCheckStatusRejected
...
  domain.CreateDriverRequest:
    properties:
      birth_date:
        type: string
      email:
        type: string
      first_name:
        maxLength: 255
        minLength: 2
        type: string
      id:
        maxLength: 20
        minLength: 3
        type: string
      last_name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        maxLength: 50
        minLength: 8
        type: string
      photo_url:
        type: string
      rut_or_dni:
        maxLength: 50
        minLength: 8
        type: string
      status:
        $ref: '#/definitions/domain.DriverStatus'
    required:
    - first_name
    - id
    - last_name
    - rut_or_dni
    - status
    type: object
...
  domain.Driver:
    properties:
      availability:
        $ref: '#/definitions/domain.DriverAvailability'
      background_check:
        $ref: '#/definitions/domain.DriverBackgroundCheck'
      birth_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      kpis:
        $ref: '#/definitions/domain.DriverKPIs'
      last_name:
        type: string
      license:
        allOf:
        - $ref: '#/definitions/domain.DriverLicense'
        description: Related data
      phone:
        type: string
      photo_url:
        type: string
      rut_or_dni:
        type: string
      status:
        $ref: '#/definitions/domain.DriverStatus'
      updated_at:
        type: string
      vehicle:
        $ref: '#/definitions/domain.Vehicle'
    type: object
  domain.DriverAvailability:
    properties:
      days:
        items:
          type: string
        type: array
      driver_id:
        type: string
      regions:
        items:
          type: string
        type: array
      time_ranges:
        items:
          $ref: '#/definitions/domain.TimeRange'
        type: array
      updated_at:
        type: string
    type: object
  domain.DriverBackgroundCheck:
    properties:
      checked_at:
        type: string
      driver_id:
        type: string
      file_url:
        type: string
      status:
        $ref: '#/definitions/domain.BackgroundCheckStatus'
    type: object
...
  domain.DriverKPIs:
    properties:
      average_rating:
        type: number
      cancel_rate:
        type: number
      on_time_rate:
        type: number
      total_km:
        type: number
      total_trips:
        type: integer
    type: object
  domain.DriverLicense:
    properties:
      class:
        $ref: '#/definitions/domain.LicenseClass'
      driver_id:
        type: string
      expires_at:
        type: string
      file_url:
        type: string
      issued_at:
        type: string
      number:
        type: string
    type: object
  domain.DriverStatus:
    enum:
    - ACTIVE
    - INACTIVE
    type: string
    x-enum-varnames:
    - DriverStatusActive
    - DriverStatusInactive
  domain.LicenseClass:
    enum:
    - A1
    - A2
    - A3
    - A4
    - A5
    - B
    - C
    - D
    - E
    type: string
    x-enum-varnames:
    - LicenseClassA1
    - LicenseClassA2
    - LicenseClassA3
    - LicenseClassA4
    - LicenseClassA5
    - LicenseClassB
    - LicenseClassC
    - LicenseClassD
    - LicenseClassE
...
  domain.TimeRange:
    properties:
      from:
        type: string
      to:
        type: string
    type: object
...
  domain.UpdateDriverRequest:
    properties:
      birth_date:
        type: string
      email:
        type: string
      first_name:
        maxLength: 255
        minLength: 2
        type: string
      last_name:
        maxLength: 255
        minLength: 2
        type: string
      phone:
        maxLength: 50
        minLength: 8
        type: string
      photo_url:
        type: string
      rut_or_dni:
        maxLength: 50
        minLength: 8
        type: string
      status:
        $ref: '#/definitions/domain.DriverStatus'
    type: object
...
  domain.Vehicle:
    properties:
      brand:
        type: string
      color:
        type: string
      created_at:
        type: string
      driver_id:
        type: string
      id:
        type: string
      inspection_expires_at:
        type: string
      insurance_expires_at:
        type: string
      insurance_policy:
        type: string
      model:
        type: string
      photos:
        items:
          type: string
        type: array
      plate:
        type: string
      type:
        $ref: '#/definitions/domain.VehicleType'
      updated_at:
        type: string
      vin:
        type: string
      year:
        type: integer
    type: object
  domain.VehicleType:
    enum:
    - BUS
    - VAN
    - SEDAN
    - SUV
    type: string
    x-enum-varnames:
    - VehicleTypeBus
    - VehicleTypeVan
    - VehicleTypeSedan
    - VehicleTypeSUV
...
  handler.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        type: string
    type: object
  handler.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@turivo.com
    name: API Support
    url: http://www.turivo.com/support
  description: API backend for Turivo transportation management system
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Turivo Backend API
  version: "1.0"
paths:
...
  /api/v1/drivers:
    get:
      consumes:
      - application/json
      description: Get paginated list of drivers
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      - description: Filter by status
        enum:
        - ACTIVE
        - INACTIVE
        in: query
        name: status
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Sort by field
        enum:
        - name
        - id
        - created_at
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Driver'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List drivers
      tags:
      - drivers
    post:
      consumes:
      - application/json
      description: Create a new driver
      parameters:
      - description: Driver data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CreateDriverRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create driver
      tags:
      - drivers
  /api/v1/drivers/{id}:
    delete:
      consumes:
      - application/json
      description: Delete driver by ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete driver
      tags:
      - drivers
    get:
      consumes:
      - application/json
      description: Get driver by ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get driver
      tags:
      - drivers
    patch:
      consumes:
      - application/json
      description: Update driver by ID
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      - description: Driver data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateDriverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Driver'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update driver
      tags:
      - drivers
  /api/v1/drivers/{id}/kpis:
    get:
      consumes:
      - application/json
      description: Get driver key performance indicators
      parameters:
      - description: Driver ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DriverKPIs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get driver KPIs
      tags:
      - drivers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"